VulkanSample: select 3d object by picking pixel

+++ Immortal/sl/slcast.h
@@ -83,7 +83,7 @@ inline constexpr uint32_t U32(T num)
 }
 
 template <class T>
-inline constexpr uint32_t U64(T num)
+inline constexpr uint64_t U64(T num)
 {
     return static_cast<uint64_t>(num);
 }
